Remove the executable property to allow override.
This problem was discussed here : SublimeLinter/SublimeLinter#455

If the `executable` property is defined, the plugin require the host system to have a global `phpcs` binary. If I haven't that binary installed (eg. I use composer to install inside my project folder) and I use the `*.sublime-project` file to configure my linter, the phpcs linter is never executed.

```
{
    "folders":
    [
        {
            "path": "."
        }
    ],
    "SublimeLinter": {
        "linters": {
            "phpcs": {
                "standard": "${folder}/phpcs.xml",
                "cmd": "${folder}/vendor/bin/phpcs"
            }
        }
    }
}
```

With that update suggested by @kaste the global binary is returned only if the configuration doesn't defined a specific one.